Change data type of tensors in unit tests, add casting in visualization pre-processing func

+++ tests/test_visualizer.py
@@ -1,6 +1,5 @@
 # pylint: disable = missing-class-docstring
 
-import numpy as np
 import pytest
 import torch
 from matplotlib.pyplot import Figure
@@ -12,8 +11,8 @@ def test_single_channel_visualization_selected_channel_should_raise_exception() 
     """Test if function raises ValueError when passing selected_channel
     not in range of color dimension of attributions matrix."""
     selected_channel: int = 4
-    attributions: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
-    transformed_img: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
+    attributions: torch.Tensor = torch.zeros((3, 2, 2), dtype=torch.float)
+    transformed_img: torch.Tensor = torch.zeros((3, 2, 2), dtype=torch.float)
     with pytest.raises(ValueError):
         _ = single_channel_visualization(
             attributions=attributions,
@@ -26,8 +25,8 @@ def test_single_channel_visualization_selected_channel_should_pass() -> None:
     """Test if function returns Figure object when passing selected_channel
     in range of color dimension of attributions matrix."""
     selected_channel: int = 0
-    attributions: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
-    transformed_img: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
+    attributions: torch.Tensor = torch.zeros((3, 2, 2), dtype=torch.float)
+    transformed_img: torch.Tensor = torch.zeros((3, 2, 2), dtype=torch.float)
     result = single_channel_visualization(
         attributions=attributions,
         transformed_img=transformed_img,
